samuel_clemens/pyproject.toml
[project]
name = "samuel-clemens"
version = "0.1.0"
description = "The Complete Methods Package for MarkTwainVerse - Frontier storytelling, narrative generation, and NSPFRP protocol integration"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "MarkTwainVerse Team" }
]
keywords = ["marktwainverse", "syntheverse", "nspfrp", "storytelling", "frontier"]

dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "Pillow>=10.0",
    "reportlab>=4.0",
]

[project.scripts]
twain = "samuel_clemens.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/samuel_clemens"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
